improv: game session error UX
Now we handle the different errors that creating a game session can
return, and inform the user accordingly so they know what to do next.

Particularly with the "new user has no wallet" case, we inform them
and guide them to make a deposit instead of shoving the deposit modal
in their face without context, and when they close it they see a broken
iframe with the text "E_INSUFFICIENT_FUNDS".

This also sets the ground for playing demo games, which will be handled
in a similar way.