feat: move KYC policy to nuxt content module in MD
Tried following the i18n integration guide in the nuxt content docs
(https://content.nuxt.com/docs/integrations/i18n), but because our
locales have a country code (en-US, pt-BR), nuxt content gave an error
saying the collection was invalid.

I wanted to keep all of the content separated by folders, as it makes
it way easier to manage, in case of adding a new locale or detecting
missing files. So the next simplest solution was doing only one
collection "pages" and separating the content files into locale
subfolders (en-US/, pt-BR/). This will become a problem if we want to
add more content types in the future aside from pages, but for now it
works fine.

Also, including the `better-sqlite3` package because it is required by
nuxt/content for some reason.